Refaktoroi AineistoTable.
