+++ src/components/projekti/common/Aineistot/AineistoTable/index.ts
@@ -1,0 +1,4 @@
+export * from "./ActionsColumn";
+export * from "./AineistoTable";
+export * from "./AineistoAlakategoria";
+export * from "./AineistoPaakategoriaContent";
